Move MQTT related logic to a sub-package
